feat: enable ext communication helper (#10)
## related KILTprotocol/ticket/issues/2126

This PR adds a helper to signal the extension that the DApp supports the
credential API.

The Credential API version is hard coded inside this library as it is
expected that downstream projects support exactly the version that is
supported by this library. If this library updates the Credential API
version all downstream project will be required to update their code to
support the new version.

### Support for possible other APIs

If there are new APIs added to the wallets this method can be extended
in the following way:

```ts
export function initializeKiltExtensionAPI(options: {
  credentialApi: boolean,
  newApi: boolean,
} = {
  credentialApi: true,
  newApi: false,
})
```

This would be a non breaking change.